# demo-backend/openapi/all_linear_scenarios.yaml
openapi: 3.0.3
info:
  title: Demo Backend - All Linear Scenarios
  version: 1.0.0
  description: |
    Lightweight demo API with only linear workflows.

    Travel workflow:
    1) GET /users/recent
    2) GET /hotels/top
    3) POST /segments/hotel
    4) POST /assignments/hotels
    5) POST /emails/send-offers

    CRM workflow:
    1) GET /crm/leads/recent
    2) POST /crm/leads/qualify
    3) POST /crm/offers/prepare
    4) POST /crm/offers/send
servers:
  - url: http://demo-api:8010
  - url: http://localhost:8010
paths:
  /users/recent:
    get:
      operationId: getRecentUsers
      tags: [travel-offer-workflow]
      summary: Get recent users
      parameters:
        - in: query
          name: last_active_after
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 30
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecentUsersResponse"

  /hotels/top:
    get:
      operationId: getTopHotels
      tags: [travel-offer-workflow]
      summary: Get top hotels
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
        - in: query
          name: city
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Hotels list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopHotelsResponse"

  /segments/hotel:
    post:
      operationId: segmentUsersByHotelPreferences
      tags: [travel-offer-workflow]
      summary: Build hotel segments from users and hotels
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HotelSegmentsRequest"
      responses:
        "200":
          description: Segments payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HotelSegmentsResponse"

  /assignments/hotels:
    post:
      operationId: assignUsersToHotels
      tags: [travel-offer-workflow]
      summary: Build user-hotel assignments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentsRequest"
      responses:
        "200":
          description: Assignments payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentsResponse"

  /emails/send-offers:
    post:
      operationId: sendHotelOffersByEmail
      tags: [travel-offer-workflow]
      summary: Send travel offers by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailOfferRequest"
      responses:
        "200":
          description: Send summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailOfferResponse"

  /crm/leads/recent:
    get:
      operationId: getRecentLeads
      tags: [crm-linear-workflow]
      summary: Get recent leads
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - in: query
          name: source
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Leads list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecentLeadsResponse"

  /crm/leads/qualify:
    post:
      operationId: qualifyLeadsForOffer
      tags: [crm-linear-workflow]
      summary: Qualify leads
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QualifyLeadsRequest"
      responses:
        "200":
          description: Qualified leads payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QualifyLeadsResponse"

  /crm/offers/prepare:
    post:
      operationId: prepareOffersForLeads
      tags: [crm-linear-workflow]
      summary: Prepare offers from qualified leads
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrepareOffersRequest"
      responses:
        "200":
          description: Prepared offers payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrepareOffersResponse"

  /crm/offers/send:
    post:
      operationId: sendPreparedOffers
      tags: [crm-linear-workflow]
      summary: Send prepared offers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOffersRequest"
      responses:
        "200":
          description: Send summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendOffersResponse"

components:
  schemas:
    User:
      type: object
      required: [id, email, last_active]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        last_active:
          type: string
          format: date-time

    Hotel:
      type: object
      required: [id, name, city]
      properties:
        id:
          type: string
        name:
          type: string
        city:
          type: string

    Segment:
      type: object
      required: [segment_id, hotel_id, user_ids]
      properties:
        segment_id:
          type: string
        hotel_id:
          type: string
        user_ids:
          type: array
          items:
            type: string

    Assignment:
      type: object
      required: [user_id, hotel_id]
      properties:
        user_id:
          type: string
        hotel_id:
          type: string

    RecentUsersResponse:
      type: object
      required: [users]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"

    TopHotelsResponse:
      type: object
      required: [hotels]
      properties:
        hotels:
          type: array
          items:
            $ref: "#/components/schemas/Hotel"

    HotelSegmentsRequest:
      type: object
      required: [users, hotels]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        hotels:
          type: array
          items:
            $ref: "#/components/schemas/Hotel"

    HotelSegmentsResponse:
      type: object
      required: [segments]
      properties:
        segments:
          type: array
          items:
            $ref: "#/components/schemas/Segment"

    AssignmentsRequest:
      type: object
      required: [segments]
      properties:
        segments:
          type: array
          items:
            $ref: "#/components/schemas/Segment"

    AssignmentsResponse:
      type: object
      required: [assignments]
      properties:
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/Assignment"

    EmailOfferRequest:
      type: object
      required: [template_id, assignments]
      properties:
        template_id:
          type: string
          default: offer_template_2026
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/Assignment"

    FailedDelivery:
      type: object
      required: [user_id, reason]
      properties:
        user_id:
          type: string
        reason:
          type: string

    EmailOfferResponse:
      type: object
      required: [sent_count, failed_count, failed]
      properties:
        sent_count:
          type: integer
        failed_count:
          type: integer
        failed:
          type: array
          items:
            $ref: "#/components/schemas/FailedDelivery"

    Lead:
      type: object
      required: [lead_id, email, source]
      properties:
        lead_id:
          type: string
        email:
          type: string
          format: email
        source:
          type: string

    QualifiedLead:
      type: object
      required: [lead_id, email, score, tier]
      properties:
        lead_id:
          type: string
        email:
          type: string
          format: email
        score:
          type: integer
        tier:
          type: string

    PreparedOffer:
      type: object
      required: [offer_id, lead_id, channel, message]
      properties:
        offer_id:
          type: string
        lead_id:
          type: string
        channel:
          type: string
        message:
          type: string

    RecentLeadsResponse:
      type: object
      required: [leads]
      properties:
        leads:
          type: array
          items:
            $ref: "#/components/schemas/Lead"

    QualifyLeadsRequest:
      type: object
      required: [leads]
      properties:
        leads:
          type: array
          items:
            $ref: "#/components/schemas/Lead"

    QualifyLeadsResponse:
      type: object
      required: [qualified_leads]
      properties:
        qualified_leads:
          type: array
          items:
            $ref: "#/components/schemas/QualifiedLead"

    PrepareOffersRequest:
      type: object
      required: [qualified_leads]
      properties:
        qualified_leads:
          type: array
          items:
            $ref: "#/components/schemas/QualifiedLead"

    PrepareOffersResponse:
      type: object
      required: [offers]
      properties:
        offers:
          type: array
          items:
            $ref: "#/components/schemas/PreparedOffer"

    SendOffersRequest:
      type: object
      required: [offers]
      properties:
        offers:
          type: array
          items:
            $ref: "#/components/schemas/PreparedOffer"

    FailedLeadDelivery:
      type: object
      required: [lead_id, reason]
      properties:
        lead_id:
          type: string
        reason:
          type: string

    SendOffersResponse:
      type: object
      required: [sent_count, failed_count, failed]
      properties:
        sent_count:
          type: integer
        failed_count:
          type: integer
        failed:
          type: array
          items:
            $ref: "#/components/schemas/FailedLeadDelivery"